+++ server/handlers/utils/constants.js
@@ -1,9 +1,29 @@
 // block specific req number
-// export const BLACK_LIST = ['183778BR', '183779BR', '185156BR', '185157BR', '185158BR', '185161BR', '185162BR', '185163BR', '185164BR', '185165BR', '185166BR', '183784BR', '183785BR', '183786BR', '183787BR', '183788BR', '184340BR', '184341BR', '184343BR', '184344BR', '184345BR', '184346BR', '184378BR'];
-export const BLACK_LIST = ['183787BR', '183788BR', '183789BR', '184340BR', '184341BR', '184343BR', '184344BR', '184345BR', '184346BR','188212BR', '188213BR', '188214BR', '188215BR', '188216BR', '188313BR', '188314BR', '188312BR', '188315BR', '188212BR', '188213BR', '188214BR', '188215BR', '188216BR', '188420BR'];
+export const BLACK_LIST = ['183787BR', '183788BR', '183789BR', '184340BR', '184341BR', '184343BR', '184344BR', '184345BR', '184346BR', '188212BR', '188213BR', '188214BR', '188215BR', '188216BR', '188313BR', '188314BR', '188312BR', '188315BR', '188212BR', '188213BR', '188214BR', '188215BR', '188216BR', '188420BR'];
 
 export const GROUP_MAP = {
   'IME DEV': 'IME Dev',
   'IME Sustaining': 'IME Sus',
   'One FS & Data Services': 'Data Management'
 };
+
+export const EXPECT_OFFER_TREND = {
+  isilon: [0, 0, 0, 0, 1, 2, 3, 4, 6, 8, 9, 10, 11, 12, 13, 14, 16, 18, 22, 23, 24, 25, 26, 27, 29, 31, 33, 35, 37, 39, 41, 43],
+  ecs: [0, 0, 0, 0, 1, 2, 3, 4, 6, 8, 10, 12, 14, 16, 19, 22, 24, 26, 29, 30, 30, 31, 31, 32, 32, 33, 34, 35, 35, 36, 37, 38, 39, 40, 40, 41, 41, 42, 42, 43, 43, 44, 45, 46, 46],
+  overview: [0, 0, 0, 0, 2, 4, 6, 8, 12, 16, 19, 22, 25, 28, 32, 36, 40, 44, 51, 53, 54, 56, 57, 59, 61, 64, 67, 70, 72, 75, 78, 81, 39, 40, 40, 41, 41, 42, 42, 43, 43, 44, 45, 46, 46]
+}
+
+export const DATE_RANGE = {
+  isilon: {
+    START: new Date('2017/6/25'),
+    END: new Date('2018/1/28')
+  },
+  ecs: {
+    START: new Date('2017/6/25'),
+    END: new Date('2018/4/29')
+  },
+  overview: {
+    START: new Date('2017/6/25'),
+    END: new Date('2018/4/29')
+  }
+}
